fix(comparateur): corrige mauvais layout des questions principales
