+++ site/source/components/Simulateur/ListeQuestions.tsx
@@ -4,7 +4,7 @@ import * as R from 'effect/Record'
 import { useTranslation } from 'react-i18next'
 import { styled } from 'styled-components'
 
-import { Body, Button, EditIcon, Li, Ul } from '@/design-system'
+import { Body, Button, EditIcon, Li, ReturnButton, Ul } from '@/design-system'
 import { Situation } from '@/domaine/Situation'
 import { GroupeDeQuestions } from '@/hooks/useQuestionsEditorialisees'
 
@@ -13,7 +13,6 @@ import Value from '../EngineValue/Value'
 import ScrollToElement from '../utils/Scroll/ScrollToElement'
 import { useAutoScrollToQuestions } from './AutoScrollToQuestions'
 import { BoutonReset } from './BoutonReset'
-import { BoutonRetour } from './BoutonRetour'
 
 type Props<S extends Situation = Situation> = {
 	groupesDeQuestions: Record<string, GroupeDeQuestions<S>>
@@ -32,7 +31,7 @@ export const ListeQuestions = ({
 	return (
 		<ScrollToElement when={autoScrollToQuestions}>
 			<BoutonsContainer>
-				<BoutonRetour onPress={retour} />
+				<ReturnButton size="XXS" onPress={retour} />
 				<BoutonReset />
 			</BoutonsContainer>
