+++ site/source/design-system/buttons/index.ts
@@ -1,2 +1,3 @@
-export * from './Button'
+export { Button } from './Button'
 export { CloseButton } from './CloseButton'
+export { ReturnButton } from './ReturnButton'
