+++ site/source/locales/ui-en.yaml
@@ -301,7 +301,6 @@ components:
         questions-principales:
           bouton: Specify your situation
           texte: For a better simulation, you can specify your situation.
-        retour: Back
         retour-liste: Back to list
         réinitialiser: Default values
         titre: Situation
@@ -396,6 +395,7 @@ global:
     septembre: September
   non: No
   oui: Yes
+  retour: Back
 gérer:
   ressources:
     annuaire-entreprises:
@@ -1658,7 +1658,6 @@ pages:
           info-title: Benefit from Acre
           label: Activate Acre in simulation
           title: Acre
-        confirm-label: Save options
         imposition:
           body: You can <2>choose between corporate and income taxation</2> for the first
             5 years.
