+++ site/source/locales/ui-fr.yaml
@@ -316,7 +316,6 @@ components:
         questions-principales:
           bouton: Préciser votre situation
           texte: Pour une meilleure simulation, vous pouvez préciser votre situation.
-        retour: Retour
         retour-liste: Revenir à la liste
         réinitialiser: Valeurs par défaut
         titre: Situation
@@ -413,6 +412,7 @@ global:
     septembre: septembre
   non: Non
   oui: Oui
+  retour: Retour
 gérer:
   ressources:
     annuaire-entreprises:
@@ -1758,7 +1758,6 @@ pages:
           info-title: Bénéficier de l’Acre
           label: Activer l’Acre dans la simulation
           title: Acre
-        confirm-label: Enregistrer les options
         imposition:
           body: Vous pouvez <2>choisir entre l’imposition sur les sociétés et sur le
             revenu</2> durant les 5 premières années.
